requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
